fix: use cluster in main
