+++ Dockerfile
@@ -32,5 +32,5 @@ RUN npx rimraf client
 
 EXPOSE 3000
 
-CMD ["yarn", "start:prod"]
+CMD ["yarn", "start:prod-cluster"]
