+++ server/package.json
@@ -10,13 +10,13 @@
     "build": "nest build && yarn copy:prod",
     "build:fix": "yarn build && yarn run:fix",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "start": "yarn copy:dev && cross-env NODE_ENV=dev nest start",
+    "start": "yarn copy:dev && cross-env NODE_ENV=dev MAX_CPU_NUM=1 nest start",
     "start:fix": "yarn build:fix && yarn start:prod",
     "start:cluster": "yarn build:fix && yarn start:prod-cluster",
     "start:dev": "yarn copy:dev && cross-env NODE_ENV=dev nest start --watch",
     "start:debug": "yarn copy:dev && cross-env NODE_ENV=dev nest start --debug --watch",
-    "start:prod": "cross-env NODE_ENV=production node -r ./scripts/tsconfig-paths dist/main",
-    "start:prod-cluster": "cross-env NODE_ENV=production node -r ./scripts/tsconfig-paths dist/cluster",
+    "start:prod": "cross-env NODE_ENV=production MAX_CPU_NUM=1 node -r ./scripts/tsconfig-paths dist/main",
+    "start:prod-cluster": "cross-env NODE_ENV=production node -r ./scripts/tsconfig-paths dist/main",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
