--- server/src/cluster.ts
@@ -1,12 +1,0 @@
-import cluster from "cluster";
-import { Master } from "./clusters";
-
-function bootstrap() {
-  if (cluster.isMaster) {
-    Master.Create(cluster);
-    return;
-  }
-  require("./main");
-}
-
-bootstrap();
